backend/requirements.txt
# AGV Web Control - Backend Dependencies

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Validation
pydantic>=2.0.0

# File handling
python-multipart>=0.0.6
Pillow>=10.0.0
PyYAML>=6.0

# WebSocket
websockets>=11.0

# ROS2 (installed via rosdep/apt)
# rclpy
# geometry_msgs
# nav_msgs
# sensor_msgs
# nav2_msgs
